api/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
google-genai
httpx
python-multipart
python-dotenv
